试卷列表和样式调整, 参数验证补充
